refactor test host to common
