+++ packages/flutter_tools/test/general.shard/resident_web_runner_test.dart
@@ -33,7 +33,6 @@ import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 import '../src/common.dart';
 import '../src/context.dart';
 import '../src/testbed.dart';
-import 'vmservice_test.dart';
 
 const List<VmServiceExpectation> kAttachLogExpectations = <VmServiceExpectation>[
   FakeVmServiceRequest(
